fix: close stale TUI clients before updates
